videoflix/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
click==8.0.4
colorama==0.4.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.2
django-debug-toolbar==3.2.4
django-import-export==2.7.1
django-redis==5.2.0
django-rq==2.5.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pycodestyle==2.8.0
pyparsing==3.0.7
PyYAML==6.0
redis==4.1.4
rq==1.10.1
sqlparse==0.4.2
tablib==3.2.0
toml==0.10.2
tzdata==2021.5
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
